Move DTOs to their modules
