+++ applications/backend/src/modules/auth/dto/auth.dto.ts
