+++ applications/backend/src/modules/auth/dto/register.dto.ts
@@ -1,0 +1,4 @@
+import { IntersectionType } from '@nestjs/swagger';
+import { CreateUserDto } from '../../users/dto/create-user.dto';
+
+export class RegisterDto extends IntersectionType(CreateUserDto) {}
