+++ applications/backend/src/modules/projects-to-users/dto/create-project-to-user.dto.ts
@@ -1,5 +1,5 @@
-import { UserEntity } from '../../entities/user.entity';
-import { ProjectEntity } from '../../entities/project.entity';
+import { UserEntity } from '../../../entities/user.entity';
+import { ProjectEntity } from '../../../entities/project.entity';
 
 export class CreateProjectToUserDto {
   isOwner?: boolean;
