+++ applications/backend/src/modules/users/dto/create-user.dto.ts
