+++ applications/backend/src/modules/users/dto/update-user.dto.ts
@@ -1,4 +1,4 @@
-import { UserEntity } from '../../entities/user.entity';
+import { UserEntity } from '../../../entities/user.entity';
 
 export class UpdateUserDto {
   id: string;
